+++ CJBItemSpawner/release-notes.md
@@ -4,6 +4,7 @@
 ## Upcoming release
 * Fixed spawning new flooring and rings (thanks to Mizzion!).
 * Fixed spawning custom rings added by mods.
+* Fixed errors when some item data is invalid.
 
 ## 1.11.5
 Released 26 November 2019 for SMAPI 3.0 or later. Updated by Pathoschild.